BWS-163: * more suggestions + loader on table refresh
