+++ frontend/src/helpers/helpers.ts
@@ -6,4 +6,4 @@ export { getQueryString } from './http/http';
 export { getDateDecoratedWithAgo } from './date/date';
 export { checkHasPermission } from './permission/permission';
 export { debounce } from './timeout/timeout';
-export { getPrettyBytes } from 'helpers/string/string.helper';
+export { getPrettyBytes } from './string/string';
