+++ frontend/src/helpers/string/string.ts
@@ -1,0 +1,1 @@
+export { getPrettyBytes } from './pretty-bytes/pretty-bytes.helper';
