+++ frontend/src/services/saver/saver.service.ts
@@ -10,6 +10,18 @@ class Saver {
     });
     saveAs(csvFile);
   }
+
+  public saveBlob(blob: Blob, filename: string): void {
+    const url = window.URL.createObjectURL(blob);
+    const a = document.createElement('a');
+    a.style.display = 'none';
+    a.href = url;
+    a.download = filename;
+    document.body.appendChild(a);
+    a.click();
+    document.body.removeChild(a);
+    window.URL.revokeObjectURL(url);
+  }
 }
 
 export { Saver };
